pyproject.toml
[project]
name = "parsing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "drissionpage>=4.1.0.17",
    "gspread>=6.2.0",
    "loguru>=0.7.3",
    "mypy>=1.15.0",
    "pydantic>=2.10.6",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "python-dotenv>=1.0.1",
    "telebot>=0.0.5",
]

[tool.black]
line-length = 79  # Ограничение длины строки (PEP 8)
target-version = ['py311', 'py312']  # Версия Python
skip-magic-trailing-comma = true  # Отключает автоматический перенос при длинных списках
